backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic[email]==2.9.2
email-validator==2.2.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
alembic==1.14.0
httpx==0.27.2
anthropic==0.34.2
python-dotenv==1.0.1
pyyaml==6.0.2
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
